pyproject.toml
[project]
name = "claude-ui"
version = "0.1.0"
description = "A sophisticated web frontend for managing Claude Code instances"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "claude-code-sdk>=0.0.19",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.0.0",
    "httpx>=0.27.0",
    "aiojobs>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "alembic>=1.13.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "black>=24.0.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
claude-ui = "claude_ui.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["claude_ui"]
